remove redundant bookinfo ambient tests and add waypoint template change test o baseline tests
